requirements.txt
PyYAML==5.3.1
coloredlogs==14.0
jsonschema==3.2.0
memoization==0.3.1
pymunk==5.6.0
pytest==6.0.1
python-socketio==4.6.0
requests==2.24.0
torch==1.6.0
tqdm==4.48.2
websocket-client==0.57.0
